+++ src/QuizContext.js
@@ -17,42 +17,21 @@ export const QuizContext = React.createContext(
     }
 );
 export function initContextValue() {
-    const defaultQuestion = {
-        id: "1001",
-        answers: [{body: "A",
-                    id: 0,
-                    answer_count: 124
-                },
-                {
-                    body: "B",
-                    id: 1,
-                    answer_count: 201
-                },
-                {
-                    body: "D",
-                    id: 2,
-                    answer_count: 150
-        }],
-        title: "Pytanie?",
-        seconds_left: 10,
-        status: "finished",
-        correct_answer: 1
-    };
     const [registeredUser, setRegisteredUser] = useState("");
     const [questionNumber, setQuestionNumber] = useState(1);
     const [quizStatus, setQuizStatus] = useState("scheduled");
     const [youTubeId, setYouTubeId] = useState("OQGtryhPZ3c-");
-    const [question, setQuestion] = useState(defaultQuestion);
+    const [question, setQuestion] = useState({});
     const [totalPlayers, setTotalPlayers] = useState(1);
     const [socket, setSocket] = useState(null);
     const [chosenAnswerId, setChosenAnswerId] = useState("");
     const [questionCount, setQuestionCount] = useState("10");
     const [player, setPlayer] = useState({});
 
-    useEffect(()=>{
+    useEffect(() => {
         const socket = io('wss://schibspy-server.herokuapp.com');
         setSocket(socket);
-        socket.on('quiz', function(msg){
+        socket.on('quiz', (msg) => {
             if(msg.current_question) {
                 setQuestion(msg.current_question);
             }
